# api/openapi/gateway.v0.yaml
openapi: 3.0.3
info:
  title: Arda Web Gateway API
  version: 0.1.0
  description: >
    MVP HTTP contract for Arda Web Gateway. WebSocket remains primary for live terminal events.
servers:
  - url: http://localhost:8080
paths:
  /api/v0/ws/terminal:
    get:
      operationId: getTerminalWebsocket
      summary: Upgrade to terminal WebSocket stream
      description: >
        Performs HTTP Upgrade to WebSocket. Runtime events carry session lifecycle and
        terminal output payloads using schemas under `components.schemas`.
        First message after connect is `session.status` so the frontend can hydrate queue/session metadata.
      responses:
        "101":
          description: Switching Protocols (WebSocket upgrade successful)
        "400":
          $ref: "#/components/responses/ErrorEnvelope"
        default:
          $ref: "#/components/responses/ErrorEnvelope"
  /api/v0/health:
    get:
      operationId: getHealth
      summary: Service health and build metadata
      responses:
        "200":
          description: Healthy response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessHealthResponse"
  /api/v0/session/connect:
    post:
      operationId: postSessionConnect
      summary: Start gateway-to-MUD session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionConnectRequest"
      responses:
        "200":
          description: Session connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSessionStatusResponse"
        "400":
          $ref: "#/components/responses/InvalidRequestError"
        "409":
          $ref: "#/components/responses/SessionAlreadyConnectedError"
        "502":
          $ref: "#/components/responses/UpstreamUnavailableError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v0/session/disconnect:
    post:
      operationId: postSessionDisconnect
      summary: Gracefully stop session
      responses:
        "200":
          description: Session disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSessionStatusResponse"
        "409":
          $ref: "#/components/responses/SessionNotConnectedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v0/session/status:
    get:
      operationId: getSessionStatus
      summary: Current session status and queue depth
      responses:
        "200":
          description: Current status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSessionStatusResponse"
  /api/v0/state/snapshot:
    get:
      operationId: getStateSnapshot
      summary: Latest parser-derived state projection
      responses:
        "200":
          description: State snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessStateSnapshotResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v0/suggestions/latest:
    get:
      operationId: getSuggestionsLatest
      summary: Latest suggestion payload and runtime status
      responses:
        "200":
          description: Latest suggestion payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSuggestionResponse"
  /api/v0/suggestions/cancel:
    post:
      operationId: postSuggestionsCancel
      summary: Cancel all in-flight suggestion requests
      responses:
        "200":
          description: Cancel request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessSuggestionCancelResponse"
  /api/v0/commands/enqueue:
    post:
      operationId: postCommandsEnqueue
      summary: Enqueue user command
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnqueueCommandRequest"
      responses:
        "200":
          description: Accepted/rejected enqueue result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnqueueResponse"
        default:
          $ref: "#/components/responses/ErrorEnvelope"
components:
  responses:
    ErrorEnvelope:
      description: Error response envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InvalidRequestError:
      description: Invalid request body or required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    SessionNotConnectedError:
      description: Session operation requires an active session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    SessionAlreadyConnectedError:
      description: Session is already connected
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    QueueFullError:
      description: Command queue is full
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UpstreamUnavailableError:
      description: Upstream MUD host is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal gateway failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    EmptyObject:
      type: object
      additionalProperties: false
    Meta:
      type: object
      required: [request_id, timestamp]
      properties:
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - INVALID_REQUEST
            - SESSION_NOT_CONNECTED
            - SESSION_ALREADY_CONNECTED
            - QUEUE_FULL
            - UPSTREAM_UNAVAILABLE
            - INTERNAL_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error, meta]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        meta:
          $ref: "#/components/schemas/Meta"
    HealthPayload:
      type: object
      required: [status, version, commit]
      properties:
        status:
          type: string
        version:
          type: string
        commit:
          type: string
    SessionConnectRequest:
      type: object
      required: [host, port]
      properties:
        host:
          type: string
        port:
          type: integer
        credential_alias:
          type: string
    SessionStatusPayload:
      type: object
      required: [session_id, connected, queue_depth, queue_max]
      properties:
        session_id:
          type: string
        connected:
          type: boolean
        queue_depth:
          type: integer
        queue_max:
          type: integer
        last_activity_at:
          type: string
          format: date-time
    SuggestionPayload:
      type: object
      required: [commands, reason, expected_outcome, generated_at]
      properties:
        commands:
          type: array
          items:
            type: string
        reason:
          type: string
        expected_outcome:
          type: string
        generated_at:
          type: string
          format: date-time
    SuggestionLatestPayload:
      type: object
      required: [in_progress]
      properties:
        in_progress:
          type: boolean
        last_error:
          type: string
        last_error_at:
          type: string
          format: date-time
        commands:
          type: array
          items:
            type: string
        reason:
          type: string
        expected_outcome:
          type: string
        generated_at:
          type: string
          format: date-time
    SuggestionCancelPayload:
      type: object
      required: [canceled]
      properties:
        canceled:
          type: boolean
    StateSnapshotPayload:
      type: object
      properties:
        location:
          type: string
        prompt:
          $ref: "#/components/schemas/PromptTuplePayload"
        status_tags:
          type: array
          items:
            type: string
        equipment:
          type: object
          additionalProperties:
            type: string
        updated_at:
          type: string
          format: date-time
      example:
        location: Таверна "Гарцующий пони"
        prompt:
          hp_current: 802
          hp_max: 802
          ma_current: 619
          ma_max: 619
          mv_current: 304
          mv_max: 304
          exp: 11911196
        status_tags: ["Белая Аура", "В полете", "Светится"]
        equipment:
          head: рогатый шлем
          wield: эльфийский клинок
        updated_at: "2026-03-02T11:12:18.000000000Z"
    StateSnapshotPayloadOrEmpty:
      oneOf:
        - $ref: "#/components/schemas/StateSnapshotPayload"
        - $ref: "#/components/schemas/EmptyObject"
    PromptTuplePayload:
      type: object
      required: [hp_current, hp_max, ma_current, ma_max, mv_current, mv_max, exp]
      properties:
        hp_current:
          type: integer
        hp_max:
          type: integer
        ma_current:
          type: integer
        ma_max:
          type: integer
        mv_current:
          type: integer
        mv_max:
          type: integer
        exp:
          type: integer
    EnqueueCommandRequest:
      type: object
      required: [command]
      properties:
        command:
          type: string
    EnqueuePayload:
      type: object
      required: [accepted, queue_depth, queue_max]
      properties:
        accepted:
          type: boolean
        queue_depth:
          type: integer
        queue_max:
          type: integer
    TerminalEvent:
      oneOf:
        - $ref: "#/components/schemas/TerminalSessionStatusEvent"
        - $ref: "#/components/schemas/TerminalSessionConnectedEvent"
        - $ref: "#/components/schemas/TerminalSessionDisconnectedEvent"
        - $ref: "#/components/schemas/TerminalOutputEvent"
        - $ref: "#/components/schemas/QueueAcceptedEvent"
        - $ref: "#/components/schemas/QueueRejectedEvent"
        - $ref: "#/components/schemas/QueueSendFailedEvent"
      discriminator:
        propertyName: event
      description: Runtime WebSocket event envelope for terminal stream updates.
    TerminalSessionStatusEvent:
      type: object
      required: [event, session_id, connected, queue_depth, queue_max, timestamp]
      properties:
        event:
          type: string
          enum: [session.status]
        session_id:
          type: string
        connected:
          type: boolean
        queue_depth:
          type: integer
        queue_max:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        event: session.status
        session_id: s-1
        connected: false
        queue_depth: 0
        queue_max: 20
        timestamp: "2026-03-02T11:12:13.000000000Z"
    TerminalSessionConnectedEvent:
      type: object
      required: [event, session_id, connected, text, queue_depth, queue_max, timestamp]
      properties:
        event:
          type: string
          enum: [session.connected]
        session_id:
          type: string
        connected:
          type: boolean
        text:
          type: string
        queue_depth:
          type: integer
        queue_max:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        event: session.connected
        session_id: s-1
        connected: true
        text: connected to 86.110.194.3:7000
        queue_depth: 0
        queue_max: 20
        timestamp: "2026-03-02T11:12:14.000000000Z"
    TerminalSessionDisconnectedEvent:
      type: object
      required: [event, session_id, connected, text, queue_depth, queue_max, timestamp]
      properties:
        event:
          type: string
          enum: [session.disconnected]
        session_id:
          type: string
        connected:
          type: boolean
        text:
          type: string
        queue_depth:
          type: integer
        queue_max:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        event: session.disconnected
        session_id: s-1
        connected: false
        text: manual disconnect, dropped_unsent=2
        queue_depth: 0
        queue_max: 20
        timestamp: "2026-03-02T11:12:20.000000000Z"
    TerminalOutputEvent:
      type: object
      required: [event, session_id, text, timestamp]
      properties:
        event:
          type: string
          enum: [terminal.output]
        session_id:
          type: string
        text:
          type: string
        timestamp:
          type: string
          format: date-time
      example:
        event: terminal.output
        session_id: s-1
        text: "Exits: north east"
        timestamp: "2026-03-02T11:12:15.000000000Z"
    QueueAcceptedEvent:
      type: object
      required: [event, session_id, text, queue_depth, queue_max, timestamp]
      properties:
        event:
          type: string
          enum: [queue.accepted]
        session_id:
          type: string
        text:
          type: string
        queue_depth:
          type: integer
        queue_max:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        event: queue.accepted
        session_id: s-1
        text: look
        queue_depth: 1
        queue_max: 20
        timestamp: "2026-03-02T11:12:16.000000000Z"
    QueueRejectedEvent:
      type: object
      required: [event, session_id, text, queue_depth, queue_max, queue_reject_code, timestamp]
      properties:
        event:
          type: string
          enum: [queue.rejected]
        session_id:
          type: string
        text:
          type: string
        queue_depth:
          type: integer
        queue_max:
          type: integer
        queue_reject_code:
          type: string
          enum: [QUEUE_FULL]
        timestamp:
          type: string
          format: date-time
      example:
        event: queue.rejected
        session_id: s-1
        text: "command rejected: queue full"
        queue_depth: 20
        queue_max: 20
        queue_reject_code: QUEUE_FULL
        timestamp: "2026-03-02T11:12:17.000000000Z"
    QueueSendFailedEvent:
      type: object
      required: [event, session_id, text, queue_depth, queue_max, timestamp]
      properties:
        event:
          type: string
          enum: [queue.send_failed]
        session_id:
          type: string
        text:
          type: string
        queue_depth:
          type: integer
        queue_max:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        event: queue.send_failed
        session_id: s-1
        text: look
        queue_depth: 1
        queue_max: 20
        timestamp: "2026-03-02T11:12:18.000000000Z"
    SuccessHealthResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/HealthPayload"
        meta:
          $ref: "#/components/schemas/Meta"
    SuccessSessionStatusResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/SessionStatusPayload"
        meta:
          $ref: "#/components/schemas/Meta"
    SuccessStateSnapshotResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/StateSnapshotPayloadOrEmpty"
        meta:
          $ref: "#/components/schemas/Meta"
    SuccessSuggestionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/SuggestionLatestPayload"
        meta:
          $ref: "#/components/schemas/Meta"
    SuccessSuggestionCancelResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/SuggestionCancelPayload"
        meta:
          $ref: "#/components/schemas/Meta"
    SuccessEnqueueResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/EnqueuePayload"
        meta:
          $ref: "#/components/schemas/Meta"